+++ backend/models/user.py
@@ -1,0 +1,5 @@
+from flask_login import UserMixin
+
+
+class User(UserMixin):
+    id: int
