+++ backend/server.py
@@ -1,33 +1,83 @@
-from flask import Flask, request, jsonify, abort
+import json
+from flask import Flask, redirect, request, jsonify, abort
 from dotenv import dotenv_values
 import psycopg2
 import os
 import zon
 import bcrypt
 from flask_login import login_user, UserMixin, LoginManager
+from oauthlib.oauth2 import WebApplicationClient
+import requests
+
+from models.user import User
+from routes.login import login_bp
+from routes.signup import signup_bp
+from conn import config, conn
 
 app = Flask(__name__)
 app.secret_key = os.urandom(32)
-
-config = {
-    **dotenv_values(".env"),
-    **dotenv_values(".env.development.local"),
-    **os.environ,
-}
-conn = psycopg2.connect(
-    database=config["POSTGRES_DATABASE_NAME"],
-    host=config["POSTGRES_DATABASE_HOST"],
-    user=config["POSTGRES_DATABASE_USER"],
-    password=config["POSTGRES_DATABASE_PASSWORD"],
-    port=config["POSTGRES_DATABASE_PORT"],
-)
-
-class User(UserMixin):
-    ...
+app.register_blueprint(login_bp)
+app.register_blueprint(signup_bp)
 
 login_manager = LoginManager()
 login_manager.init_app(app)
 login_manager.session_protection = "strong"
+
+client = WebApplicationClient(config["GOOGLE_OAUTH_CLIENT_ID"])
+
+GOOGLE_DISCOVERY_URL = "https://accounts.google.com/.well-known/openid-configuration"
+
+
+def get_google_provider_cfg():
+    return requests.get(GOOGLE_DISCOVERY_URL).json()
+
+
+@app.route("/oauth/login")
+def oauth_login():
+    google_provider_cfg = get_google_provider_cfg()
+    authorization_endpoint = google_provider_cfg["authorization_endpoint"]
+
+    request_uri = client.prepare_request_uri(
+        authorization_endpoint,
+        redirect_uri=request.base_url + "/callback",
+        scope=["openid", "email", "profile"],
+    )
+    return redirect(request_uri)
+
+
+@app.route("/oauth/login/callback")
+def callback():  # https://realpython.com/flask-google-login/
+    code = request.args.get("code")
+
+    google_provider_cfg = get_google_provider_cfg()
+    token_endpoint = google_provider_cfg["token_endpoint"]
+
+    token_url, headers, body = client.prepare_token_request(
+        token_endpoint,
+        authorization_response=request.url,
+        redirect_url=request.base_url,
+        code=code,
+    )
+
+    token_response = requests.post(
+        token_url,
+        headers=headers,
+        data=body,
+        auth=(config["GOOGLE_OAUTH_CLIENT_ID"], config["GOOGLE_OAUTH_CLIENT_SECRET"]),
+    )
+
+    client.parse_request_body_response(json.dumps(token_response.json()))
+
+    userinfo_endpoint = google_provider_cfg["userinfo_endpoint"]
+
+    uri, headers, body = client.add_token(userinfo_endpoint)
+    userinfo_response = requests.get(uri, headers=headers, data=body)
+
+    if userinfo_response.json().get("email_verified"):
+        return None  # temp needs testing
+
+    return "Failed to authenticate user.", 400
+
 
 def get_user(id):
     cursor = conn.cursor()
@@ -35,99 +85,19 @@ def get_user(id):
     found_user = cursor.fetchone()
     if not found_user:
         return None
-    
+
     return found_user
+
 
 @login_manager.user_loader
 def user_loader(id: int):
     user = get_user(id)
-    if user: # check if user exists in db
+    if user:  # check if user exists in db
         user_model = User()
         user_model.id = id
         return user_model
     return None
 
-signup_schema = zon.record(
-    {
-        "display_name": zon.string(),
-        "email": zon.string().email().max(255).ends_with("@illinois.edu"),
-        "password": zon.string().max(
-            72
-        ),  # google "why is it ok to send plaintext passwords over HTTPS"
-    }
-).strict()
-
-
-@app.route("/signup", methods=["POST"])
-def signup():
-    validated_data = signup_schema.validate(request.json)
-    password = validated_data["password"]
-    password_bytes = password.encode()
-
-    salt = bcrypt.gensalt()
-    hashed_password = bcrypt.hashpw(
-        password_bytes, salt
-    )  # no need to store salt separately with bcrypt
-
-    cursor = conn.cursor()
-    cursor.execute(
-        "INSERT INTO users(display_name, email, password_hash) VALUES (%s, %s, %s) RETURNING id",
-        (
-            validated_data["display_name"],
-            validated_data["email"],
-            hashed_password.decode(),
-        ),
-    )
-    new_user_id = cursor.fetchone()[0]
-    cursor.close()
-
-    user_object = User()
-    user_object.id = new_user_id
-
-    login_user(user_object)
-
-    return jsonify({
-        "id": new_user_id,
-        "display_name": validated_data["display_name"],
-        "email": validated_data["email"],
-    })
-
-login_schema = zon.record(
-    {
-        "email": zon.string().email().max(255).ends_with("@illinois.edu"),
-        "password": zon.string().max(72)
-    }
-)
-
-@app.route("/login", methods=["POST"])
-def login():
-    validated_data = login_schema.validate(request.json)
-
-    cursor = conn.cursor()
-    cursor.execute("SELECT id, display_name, email, password_hash FROM users WHERE email = %s", (validated_data['email'],))
-    found_user = cursor.fetchone()
-    cursor.close()
-
-    if not found_user:
-        abort(401, "Invalid username or password.")
-    
-    id, display_name, email, password_hash = found_user
-    hashed_password = password_hash.encode()
-    password_bytes = validated_data['password'].encode()
-
-    if not bcrypt.checkpw(password_bytes, hashed_password):
-        abort(401, "Invalid username or password.")
-
-    user_object = User()
-    user_object.id = id
-    
-    login_user(user_object)
-
-    return jsonify( {
-        "id": id,
-        "display_name": display_name,
-        "email": email
-    })
 
 if __name__ == "__main__":
     app.run(debug=True)
